# src/docs/swagger-output.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "API مشروع بثواني",
    "description": "وثائق RESTful API لمشروع بثواني مُولَّدة أوتوماتيكيًا بواسطة swagger-autogen",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000/api/v1",
      "description": "سيرفر التطوير المحلي"
    }
  ],
...
  "paths": {
...
    "/daily-offers": {
      "get": {
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/nearby/new": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "lat",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "lng",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "apiKey",
        "name": "Authorization",
        "in": "header",
        "description": "JWT Authorization header. مثال: Bearer <token>"
      }
    },
...
  }
}